perf(agor-ui): lazy-load syntax highlighter
ThemedSyntaxHighlighter statically imported react-syntax-highlighter (Prism +
bundled grammars, ~140KB), so the highlighter loaded eagerly with every
consumer (tool-use renderers, code preview, consent modal) even before any
code block rendered.

Split the Prism render into ThemedSyntaxHighlighter.inner.tsx and pull it in
via React.lazy (mirroring CodeEditor). The fallback renders the raw code in
the caller's PreTag wrapper, so the outer element identity and layout are
preserved while the chunk downloads. The component's public API is unchanged.
